:sparkles: Alterando banco de dados para ordenar as refeições por data

+++ src/use-cases/meals/delete-meal.ts
@@ -1,13 +1,14 @@
 import { TMealsRepository } from "@/repositories/meals-repository"
-import { MealNotFoundError } from "../errors"
+import { MealNotFoundError, UnauthorizedAccessError } from "../errors"
 
 export class DeleteMeal {
   constructor(private mealsRepository: TMealsRepository) {}
 
-  async execute(mealId: string) {
+  async execute(mealId: string, userId: string) {
     const meal = await this.mealsRepository.getMealById(mealId)
 
     if (!meal) throw new MealNotFoundError()
+    if (meal.user_id !== userId) throw new UnauthorizedAccessError()
 
     await this.mealsRepository.deleteMeal(mealId)
   }
